% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_stats_tests.R
\name{sk_ks_test}
\alias{sk_ks_test}
\title{Kolmogorov-Smirnov Test of Normality}
\usage{
sk_ks_test(x, verbose = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{verbose}{Logical; if TRUE, prints detailed results}
}
\value{
List with test results
}
\description{
Performs Kolmogorov-Smirnov test for normality.
}
