% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_stats_tests.R
\name{sk_shapiro_test}
\alias{sk_shapiro_test}
\title{Shapiro-Wilk Test of Normality}
\usage{
sk_shapiro_test(x, verbose = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{verbose}{Logical; if TRUE, prints detailed results}
}
\value{
List with test results
}
\description{
Performs Shapiro-Wilk test for normality.
}
