% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_note.R
\name{sknote}
\alias{sknote}
\title{Start SKthink Console Note (sknote)}
\usage{
sknote(
  format = c("clipboard", "txt", "md", "html"),
  file_path = NULL,
  save_plots = TRUE,
  plot_dir = "sknote_plots"
)
}
\arguments{
\item{format}{Output format. Options: "clipboard", "txt", "md", "html"}

\item{file_path}{File name and location for saving}

\item{save_plots}{Logical. Whether to save figures (Default: TRUE)}

\item{plot_dir}{Directory to save figures (Default: "sknote_plots")}
}
\description{
Start SKthink Console Note (sknote)
}
